Updates to spelling, name corrections to match OLD_ASCC
